Moved dispatches to fetch actions from components to App.js reducing the number of network calls.
